ADD: roles logic

+++ src/api/main_handlers.py
@@ -5,7 +5,7 @@ from sqlalchemy.exc import IntegrityError
 from uuid import UUID
 from logging import getLogger
 from datetime import timedelta
-from src.api.handlers.users.user import user_router, _create_new_user, _delete_user, _get_user_by_id, _update_user
+from src.api.handlers.users.user import user_router, _create_new_user, _delete_user, _get_user_by_id, _update_user, check_user_permissions
 from src.api.handlers.auth.auth import login_router, authenticate_user, create_access_token, get_current_user_from_token
 from src.api.models import UserCreate, DeletedUserResponse, ShowUser, UpdatedUserResponse, UpdatedUserRequest, Token
 from src.db.session import get_db
@@ -25,6 +25,14 @@ async def delete_user(
     db: AsyncSession = Depends(get_db),
     current_user: User = Depends(get_current_user_from_token),
     ) -> DeletedUserResponse: 
+        user_to_delete = await _get_user_by_id(user_id, db)
+        if user_to_delete is None: 
+            raise HTTPException(status_code=404, detail=f"User with id {user_id} not found")
+        if not check_user_permissions(
+             target_user=user_to_delete, 
+             current_user=current_user
+        ): 
+            raise HTTPException(status_code=403, detail=f"Forbidden")             
         deleted_user_id = await _delete_user(user_id, db)
         if deleted_user_id is None: 
             raise HTTPException(status_code=404, detail=f"User with id {user_id} not found")
@@ -50,9 +58,11 @@ async def update_user(
         updated_user_params = body.model_dump(exclude_none=True)
         if updated_user_params == {}:
             raise HTTPException(status_code=422, detail="At least one parameter for user update info should be provided")
-        user = await _get_user_by_id(user_id, db)
-        if user is None:
+        user_for_update = await _get_user_by_id(user_id, db)
+        if user_for_update is None:
             raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
+        if not check_user_permissions(target_user=user_for_update, current_user=current_user):
+                raise HTTPException(status_code=403, detail="Forbidden")
         try:
             updated_user_id = await _update_user(updated_user_params=updated_user_params, session=db, user_id=user_id)
             return UpdatedUserResponse(updated_user_id=updated_user_id)
